Module Theora.Encoder

type t

type settings = {

keyframe_frequency : int option;

vp3_compatible : bool option;

soft_target : bool option;

buffer_delay : int option;

speed : int option;

}

val create : info -> settings -> (string * string) list -> t

Initialize a state handle for decoding.

val encode_header : t -> Ogg.Stream.stream -> unit

* Fills the given stream with the header packets.

val encode_page : t -> Ogg.Stream.stream -> (unit -> yuv_buffer) -> Ogg.Page.t

* Encode data until a page is filled.

val encode_buffer : t -> Ogg.Stream.stream -> yuv_buffer -> unit

Encode a buffer.

val frames_of_granulepos : t -> Stdlib.Int64.t -> Stdlib.Int64.t

Convert a granulepos to an absolute frame index, starting at 0. * The granulepos is interpreted in the context of a given theora_state handle.

val eos : t -> Ogg.Stream.stream -> unit

Set end of stream

deprecated This function generates invalid bitstream. Please use Ogg.Stream.terminate instead!